requirements.txt
awscli==1.11.136
boto3==1.4.6
botocore==1.6.3
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.7
docutils==0.14
Flask==0.12.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
line-bot-sdk==1.3.0
MarkupSafe==1.0
olefile==0.44
peewee==2.10.1
Pillow==4.2.1
psycopg2==2.7.3
pyasn1==0.3.2
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
